Module Caqti_mirage.Make

Parameters

module RANDOM : Mirage_random.S

module TIME : Mirage_time.S

module MCLOCK : Mirage_clock.MCLOCK

module PCLOCK : Mirage_clock.PCLOCK

module STACK : Tcpip.Stack.V4V6

module DNS : Dns_client_mirage.S

Signature

module Pool : Caqti_pool_sig.S with type 'a fiber := 'a Lwt.t

include Caqti_connect_sig.S with type 'a fiber := 'a Lwt.t and type ('a, 'e) stream := ('a, 'e) Caqti_lwt.Stream.t and type ('a, 'e) pool := ('a, 'e) Pool.t and type connection := Caqti_lwt.connection and type 'a with_switch := ?sw:Caqti_lwt.Switch.t -> 'a and type 'a with_stdenv := STACK.t -> DNS.t -> 'a

val connect : ?env:(Caqti_driver_info.t -> string -> Caqti_query.t) -> ?tweaks_version:(int * int) -> ?sw:Caqti_lwt.Switch.t -> STACK.t -> DNS.t -> Uri.t -> (Caqti_lwt.connection, [> Caqti_error.load_or_connect ]) Stdlib.result Lwt.t

connect uri locates and loads a driver which can handle uri, passes uri to the driver, which establish a connection and returns a first-class module implementing Caqti_connection_sig.S.

If you use preemptive threading, note that the connection must only be used from the thread where it was created.

See Database Tweaks for details about the tweaks_version parameter.

parameter tweaks_version
Declares compatibility with database tweaks introduced up to the given version of Caqti. Defaults to a conservative value. See the above reference for more info (not linked here due to odoc issue).

parameter env
If provided, this function will do a final expansion of environment variables which occurs in the query templates of the requests executed on the connection.

val with_connection : ?env:(Caqti_driver_info.t -> string -> Caqti_query.t) -> ?tweaks_version:(int * int) -> STACK.t -> DNS.t -> Uri.t -> (Caqti_lwt.connection -> ('a, [> Caqti_error.load_or_connect ] as 'e) Stdlib.result Lwt.t) -> ('a, 'e) Stdlib.result Lwt.t

with_connection uri f calls connect on uri. If connect evaluates to Ok connection, with_connection passes the connection to f. Once f either evaluates to a result, or raises an exception, with_connection closes the database connection.

parameter tweaks_version
Passed to connect.

parameter env
Passed to connect.

val connect_pool : ?pool_config:Caqti_pool_config.t -> ?post_connect: (Caqti_lwt.connection -> (unit, 'connect_error) Stdlib.result Lwt.t) -> ?env:(Caqti_driver_info.t -> string -> Caqti_query.t) -> ?tweaks_version:(int * int) -> ?sw:Caqti_lwt.Switch.t -> STACK.t -> DNS.t -> Uri.t -> ((Caqti_lwt.connection, [> Caqti_error.connect ] as 'connect_error) Pool.t, [> Caqti_error.load ]) Stdlib.result

connect_pool uri is a pool of database connections constructed by connect uri.

Do not use pooling for connections to volatile resources like sqlite3::memory: and beware of temporary tables or other objects which may not be shared across connections to the same URI.

If you use preemptive threading, note that the connection pool must only be used from the thread where it was created. Use thread local storage to create a separate pool per thread if necessary.

parameter pool_config
Provides tuning parameters for the pool. The default is the result of a fresh call of Caqti_pool_config.default_from_env.

parameter tweaks_version
Passed to connect when creating new connections.

parameter env
Passed to connect when creating new connections.